+++ CryptoNewsAndroid/app/src/main/java/com/princess/android/cryptonews/News/Activity/LatestNewsActivity.java
@@ -1,4 +1,4 @@
-package com.princess.android.cryptonews;
+package com.princess.android.cryptonews.News.Activity;
 
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
@@ -8,6 +8,8 @@ import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.view.Menu;
 import android.view.MenuItem;
+
+import com.princess.android.cryptonews.R;
 
 public class LatestNewsActivity extends AppCompatActivity {
