+++ CryptoNewsAndroid/app/src/main/java/com/princess/android/cryptonews/NewsWebsite/Test.java
@@ -1,0 +1,8 @@
+package com.princess.android.cryptonews.NewsWebsite;
+
+/**
+ * Created by numb3rs on 2/13/18.
+ */
+
+public class Test {
+}
