+++ CryptoNewsAndroid/app/src/main/res/layout/activity_latest_news.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.princess.android.cryptonews.LatestNewsActivity">
+    tools:context="com.princess.android.cryptonews.News.Activity.LatestNewsActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
