+++ CryptoNewsAndroid/app/src/main/res/layout/content_latest_news.xml
@@ -2,7 +2,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/fragment"
-    android:name="com.princess.android.cryptonews.LatestNewsActivityFragment"
+    android:name="com.princess.android.cryptonews.News.Fragment.LatestNewsActivityFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
