+++ CryptoNewsAndroid/app/src/main/res/layout/fragment_latest_news.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.princess.android.cryptonews.LatestNewsActivityFragment"
+    tools:context="com.princess.android.cryptonews.News.Fragment.LatestNewsActivityFragment"
     tools:showIn="@layout/activity_latest_news">
 
     <TextView
